backend/requirements.txt
flask==2.0.1
google-generativeai==0.3.2
python-dotenv==0.19.0
flask-cors==3.0.10
torch==2.0.1
torchvision==0.15.2
pillow==9.5.0
timm==0.9.2
numpy==1.24.3
pandas
scikit-learn
matplotlib
eli5
lime
pdpbox